+++ src/main/java/com/solutec/desarrollo_web_server/entity/RolePermissionId.java
@@ -5,8 +5,8 @@ import java.util.Objects;
 
 public class RolePermissionId implements Serializable {
 
-    private Integer permission;
-    private Integer role;
+    private Long permission;
+    private Long role;
 
     @Override
     public boolean equals(Object o) {
